pyproject.toml
[tool.pdm.build]
includes = []
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[project]
authors = [
    {name = "electimon", email = "electimon@gmail.com"},
]
requires-python = "<4.0,>=3.10"
dependencies = [
    "iso3166",
    "geopy",
    "jinja2",
    "cachetools",
    "requests",
    "yfinance",
    "feedparser",
    "fastapi",
    "langcodes",
    "starlette-context",
    "uvicorn",
    "sentry-sdk>=2.20.0",
    "timezonefinder>=6.5.8",
    "pytz>=2024.2",
    "pyephem>=9.99",
    "redis[hiredis]>=5.2.1",
]
name = "yql-x-server"
version = "0.1.0"
description = ""
readme = "README.md"
